added route for register

+++ app/routes/static.py
@@ -1,17 +1,28 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, Depends
 from fastapi.exceptions import HTTPException
 
-from models.models import Static
-from database.static import create_static, get_statc, update_static, delete_static
+from models.models import Static, User
+from utils.security import get_current_user
+from database.static import create_static, get_static, update_static, delete_static
 
 router = APIRouter()
 
 
-@router.post(
-    "/",
-    response_model=Static,
-)
-async def post_static_content(static: Static):
+@router.get("/{name}", response_model=Static)
+async def get_static_content(title: str):
+    response = await get_static(title)
+    if response:
+        return response
+    raise HTTPException(status_code=404, detail=f"Content: {title} not found!")
+
+
+@router.post("/", response_model=Static)
+async def post_static_content(
+    static: Static, user: User = Depends(get_current_user)
+) -> Static:
+    if user.role.value != "superuser":
+        raise HTTPException(status_code=403, detail="Permission denied")
+
     try:
         response = await create_static(static)
         return {"Result": response}
@@ -19,16 +30,12 @@ async def post_static_content(static: Static):
         return {"Error": error.detail}, error.status_code
 
 
-@router.get("/{name}", response_model=Static)
-async def get_static_content(title: str):
-    response = await get_statc(title)
-    if response:
-        return response
-    raise HTTPException(status_code=404, detail=f"Content: {title} not found!")
-
-
 @router.put("/{id}", response_model=Static)
-async def update_static_content(static_id: str, static: Static):
+async def update_static_content(
+    static_id: str, static: Static, user: User = Depends(get_current_user)
+) -> Static:
+    if user.role.value != "superuser":
+        raise HTTPException(status_code=403, detail="Permission denied")
     response = await update_static(static, static_id)
 
     if response:
@@ -37,7 +44,11 @@ async def update_static_content(static_id: str, static: Static):
 
 
 @router.delete("/{id}", response_model=Static)
-async def delete_static_content(static_id: str):
+async def delete_static_content(
+    static_id: str, user: User = Depends(get_current_user)
+) -> Static:
+    if user.role.value != "superuser":
+        raise HTTPException(status_code=403, detail="Permission denied")
     response = await delete_static(static_id)
 
     if response:
